Add client bot file.
