+++ zone/CMakeLists.txt
@@ -16,6 +16,7 @@ SET(zone_sources
     botspellsai.cpp
     cheat_manager.cpp
     client.cpp
+    client_bot.cpp
     client_mods.cpp
     client_packet.cpp
     client_process.cpp
